Fixed most failing tests.
RandoriBundleProjectTest had a few invalid paths.
EchoBehaviorTest and VerticalTabsTest needed to be updated with to take into account some refactoring of the Randori framework.

build.gradle was updated to use an older version of Google Guava because Falcon uses a method that was de-public-ed.  Technically Falcon uses Guava r08, but the r08 guava on the maven repository is not a valid jar, so this guava is r09